feat: #5 - PR 리뷰 반영
- 코드 컨벤션 수정
- FeatureAssembly 제거
- SceneDelegate에서 AppCoordinator 객체 의존성 등록
