+++ Targets/DesignKit/Sources/Extensions/UIViewController+.swift
@@ -29,8 +29,3 @@ public extension Reactive where Base: UIViewController {
       .map { _ in }
   }
 }
-
-public extension UIViewController {
-  static let screenHeight = UIScreen.main.bounds.height
-  static let screenWidth = UIScreen.main.bounds.width
-}
